report error back to the caller if we cannot register the dialog
we can only fall here if the path the caller expect us to use is already
used by an previous call, or different portal. since we cannot do much
here, we forward the error to the caller.

since falling to register the dialog don't cause issues to the portal daemon,
the critical is replaced by a warning.