Background: make asynchronous, get notifications proxy lazily
make all of the dbus calls and methods asynchronous so that we don't end up
blocking others interfaces, methods and initialization.

the notification proxy is now also started when we first call
send_notification() so we don't end starting it early than necessary or
blocking the initialization for too much time.

Signed-off-by: Gustavo Marques <pushstarttocontinue@outlook.com>